+++ ESM/build_wrf.csh
@@ -1,7 +1,7 @@
 #!/bin/csh -f
 #
 # git $Id$
-# svn $Id: build_wrf.csh 1066 2021-05-14 21:47:45Z arango $
+# svn $Id: build_wrf.csh 1071 2021-06-14 04:22:26Z arango $
 #::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 # Copyright (c) 2002-2021 The ROMS/TOMS Group                           :::
 #::::::::::::::::::::::::::::::::::::::::::::::::::::: Hernan G. Arango :::
@@ -102,9 +102,14 @@ end
 # ESMF/NUOPC Coupling (see below).
 #--------------------------------------------------------------------------
 
- setenv ROMS_SRC_DIR         ${HOME}/ocean/repository/coupling
+ setenv ROMS_SRC_DIR         ${HOME}/ocean/repository/trunk
 
+#setenv WRF_ROOT_DIR         ${HOME}/ocean/repository/WRF.4.1.2
+#setenv WRF_ROOT_DIR         ${HOME}/ocean/repository/WRF.4.1.3
+#setenv WRF_ROOT_DIR         ${HOME}/ocean/repository/WRF.4.2.2
+#setenv WRF_ROOT_DIR         ${HOME}/ocean/repository/WRF.4.3
  setenv WRF_ROOT_DIR         ${HOME}/ocean/repository/WRF
+
  setenv WRF_SRC_DIR          ${WRF_ROOT_DIR}
 
 # Decode WRF version from its README file to decide the appropriate data
@@ -131,13 +136,12 @@ end
 # shell's quoting syntax to enclose the definition. Both single or
 # double quotes work. For example,
 #
-#setenv MY_CPP_FLAGS         "${MY_CPP_FLAGS} -DHDF5"
+# setenv MY_CPP_FLAGS        "${MY_CPP_FLAGS} -DHDF5"
 #
-# can be used to read input data from HDF5 file instead of flat files.
 # Notice that you can have as many definitions as you want by appending
 # values.
 
- setenv MY_CPP_FLAGS         "${MY_CPP_FLAGS} -DHDF5"
+#setenv MY_CPP_FLAGS        "${MY_CPP_FLAGS} -D"
 
 #--------------------------------------------------------------------------
 # Set Fortran compiler and MPI library to use.
@@ -163,9 +167,8 @@ end
 
 #setenv USE_REAL_DOUBLE      on          # use real double precision (-r8)
 #setenv USE_DEBUG            on          # use Fortran debugging flags
- setenv USE_HDF5             on          # compile with HDF5 library
  setenv USE_NETCDF           on          # compile with NetCDF
-#setenv USE_NETCDF4          on          # compile with NetCDF-4 library
+ setenv USE_NETCDF4          on          # compile with NetCDF-4 library
                                          # (Must also set USE_NETCDF)
 
  unsetenv PNETCDF                        # disable compiling with pNetCDF
@@ -178,8 +181,8 @@ end
  setenv USE_MY_LIBS no           # use system default library paths
 #setenv USE_MY_LIBS yes          # use my customized library paths
 
-#set MY_PATHS = ${ROMS_SRC_DIR}/Compilers/my_build_paths.csh
- set MY_PATHS = ${HOME}/Compilers/ROMS/my_build_paths.csh
+ set MY_PATHS = ${ROMS_SRC_DIR}/Compilers/my_build_paths.csh
+#set MY_PATHS = ${HOME}/Compilers/ROMS/my_build_paths.csh
 
 if ($USE_MY_LIBS == 'yes') then
   source ${MY_PATHS} ${MY_PATHS}
@@ -265,63 +268,9 @@ if ( $config == 1 ) then
     setenv CONFIG_FLAGS   -r8
   endif
 
-  set CHECK_STRING = 'WRF-ROMS ESMF-NUOPC Coupling'
-  echo ""
-  echo "${separator}"
-  echo "If applicable, replacing several WRF files for WRF ESMF/NUOPC Coupling"
-  echo "${separator}"
-  echo ""
+# Check if WRF needs to be patched and do so if necessary.
 
-# Reworking linking NetCDF4 library dependencies
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/configure`) then
-    mv -v ${WRF_ROOT_DIR}/configure ${WRF_ROOT_DIR}/configure.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_configure ${WRF_ROOT_DIR}/configure
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/configure"
-  endif
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/Makefile`) then
-    mv -v  ${WRF_ROOT_DIR}/Makefile ${WRF_ROOT_DIR}/Makefile.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_Makefile  ${WRF_ROOT_DIR}/Makefile
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/Makefile"
-  endif
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/arch/postamble`) then
-    mv -v  ${WRF_ROOT_DIR}/arch/postamble ${WRF_ROOT_DIR}/arch/postamble.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_postamble ${WRF_ROOT_DIR}/arch/postamble
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/arch/postamble"
-  endif
-
-# Adding MacOS Intel/GNU with OpenMPI
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/arch/configure.defaults`) then
-    mv -v  ${WRF_ROOT_DIR}/arch/configure.defaults ${WRF_ROOT_DIR}/arch/configure.defaults.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_configure.defaults ${WRF_ROOT_DIR}/arch/configure.defaults
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/arch/configure.defaults"
-  endif
-
-# Create clean .f90 files for debugging and rename modules to WRF_ESMF_*
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/external/esmf_time_f90/Makefile`) then
-    mv -v  ${WRF_ROOT_DIR}/external/esmf_time_f90/Makefile ${WRF_ROOT_DIR}/external/esmf_time_f90/Makefile.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_Makefile.esmf ${WRF_ROOT_DIR}/external/esmf_time_f90/Makefile
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/external/esmf_time_f90/Makefile"
-  endif
-
-# Correcting optional argument from defaultCalendar to defaultCalKind in
-# ESMF_Initialize call
-
-  if (! `grep -c "${CHECK_STRING}" ${WRF_ROOT_DIR}/external/esmf_time_f90/Test1.F90`) then
-    mv -v  ${WRF_ROOT_DIR}/external/esmf_time_f90/Test1.F90 ${WRF_ROOT_DIR}/external/esmf_time_f90/Test1.F90.orig
-    cp -fv ${ROMS_SRC_DIR}/ESM/wrf_Test1.F90 ${WRF_ROOT_DIR}/external/esmf_time_f90/Test1.F90
-  else
-    echo "   No need to replace: ${WRF_ROOT_DIR}/external/esmf_time_f90/Test1.F90"
-  endif
+  ${ROMS_SRC_DIR}/ESM/wrf_patch.csh
 
   echo ""
   echo "${separator}"
@@ -331,6 +280,12 @@ if ( $config == 1 ) then
 
   ${WRF_ROOT_DIR}/configure ${CONFIG_FLAGS}
 
+# If which_MPI is "intel" then we need to replace DM_FC and DM_CC in configure.wrf
+
+  if ( "${which_MPI}" == "intel" ) then
+    perl -i -pe 's/^DM_FC(\s*)=(\s*)mpif90/DM_FC$1=$2mpiifort/' ${WRF_SRC_DIR}/configure.wrf
+    perl -i -pe 's/^DM_CC(\s*)=(\s*)mpicc/DM_CC$1=$2mpiicc/' ${WRF_SRC_DIR}/configure.wrf
+  endif
 endif
 
 #--------------------------------------------------------------------------
@@ -373,13 +328,15 @@ setenv WRF_NMM_CORE 0            # Nonhydrostatic Mesoscale Model core
 
 # Remove existing build directory.
 
-if ( $clean == 1 ) then
-  echo ""
-  echo "${separator}"
-  echo "Removing WRF build directory:  ${WRF_BUILD_DIR}"
-  echo "${separator}"
-  echo ""
-  /bin/rm -rf ${WRF_BUILD_DIR}
+if ( $move == 1 ) then
+  if ( $clean == 1 ) then
+    echo ""
+    echo "${separator}"
+    echo "Removing WRF build directory:  ${WRF_BUILD_DIR}"
+    echo "${separator}"
+    echo ""
+    /bin/rm -rf ${WRF_BUILD_DIR}
+  endif
 endif
 
 # Compile (if -move is set, the binaries will go to WRF_BIN_DIR set above).
@@ -408,229 +365,19 @@ echo ""
 ${WRF_ROOT_DIR}/compile ${WRF_CASE}
 
 #--------------------------------------------------------------------------
-# Move WRF objects and executables.
+# Move WRF objects and executables to working project directory.
 #--------------------------------------------------------------------------
 
 if ( $move == 1 ) then
-
-  echo ""
-  echo "${separator}"
-  echo "Moving WRF objects to Build directory  ${WRF_BUILD_DIR}:"
-  echo "${separator}"
-  echo ""
-
-  if ( ! -d ${WRF_BUILD_DIR} ) then
-    /bin/mkdir -pv ${WRF_BUILD_DIR}
-    /bin/mkdir -pv ${WRF_BIN_DIR}
-    echo ""
-  endif
-
-  /bin/cp -pfv configure.wrf ${WRF_BUILD_DIR}
-  /bin/cp -pfv Registry/Registry ${WRF_BUILD_DIR}
-  /bin/cp -pfv Registry/io_boilerplate_temporary.inc ${WRF_BUILD_DIR}
-
-  /bin/mv -fv run/namelist.input ${WRF_BUILD_DIR}
-  /bin/mv -fv test/em_real/README.namelist ${WRF_BUILD_DIR}
-
-  /bin/mv -fv frame/md_calls.inc ${WRF_BUILD_DIR}
-  /bin/mv -fv frame/module_dm.F ${WRF_BUILD_DIR}
-  /bin/mv -fv frame/module_state_description.F ${WRF_BUILD_DIR}
-  /bin/mv -fv external/io_grib1/io_grib1.f90 ${WRF_BUILD_DIR}
-  /bin/mv -fv external/io_grib_share/io_grib_share.f90 ${WRF_BUILD_DIR}
-
-  /bin/mv -fv tools/gen_comms.c ${WRF_BUILD_DIR}
-
-  /bin/mv -fv inc/dm_comm_cpp_flags ${WRF_BUILD_DIR}
-  /bin/mv -fv inc/wrf_io_flags.h ${WRF_BUILD_DIR}
-  /bin/mv -fv inc/wrf_status_codes.h ${WRF_BUILD_DIR}
-
-  /bin/mv -fv external/io_int/io_int_idx_tags.h ${WRF_BUILD_DIR}
-
-# The "arch", "chem",  "run" and "var" directories have source code files
-# (*.inc, *.F90, *.f90) that are part of repository. We cannot use the
-# compact "find" function for these file extensions.
-
-  foreach DIR ( frame chem share dyn_em dyn_exp dyn_nmm phys main tools wrftladj )
-    if ( -d $DIR ) then
-      /bin/mv -fv ${DIR}/*.f90 ${WRF_BUILD_DIR}
-      /bin/mv -fv ${DIR}/*.F90 ${WRF_BUILD_DIR}
-      /bin/mv -fv ${DIR}/*.inc ${WRF_BUILD_DIR}
-    endif
-  end
-
-  find ${WRF_ROOT_DIR} -type f -name "*.mod" -exec /bin/mv -fv {} ${WRF_BUILD_DIR} \;
-  find ${WRF_ROOT_DIR} -type f -name "*.o"   -exec /bin/mv -fv {} ${WRF_BUILD_DIR} \;
-  find ${WRF_ROOT_DIR} -type f -name "*.obj" -exec /bin/mv -fv {} ${WRF_BUILD_DIR} \;
-  find ${WRF_ROOT_DIR} -type f -name "*.a"   -exec /bin/mv -fv {} ${WRF_BUILD_DIR} \;
-
-  find ${WRF_ROOT_DIR} -type f -name "*.exe" -exec /bin/mv -fv {} ${WRF_BIN_DIR} \;
-  find ${WRF_ROOT_DIR} -type l -name "*.exe" -exec /bin/rm -fv {} \;
-
-  /bin/mv -fv external/esmf_time_f90/*.f ${WRF_BUILD_DIR}
-  /bin/cp -pv external/esmf_time_f90/ESMF*.inc ${WRF_BUILD_DIR}
-
-  /bin/mv -fv external/io_int/diffwrf ${WRF_BIN_DIR}/diffwrf_int
-  /bin/mv -fv external/io_int/test_io_idx ${WRF_BIN_DIR}
-  /bin/mv -fv external/io_netcdf/diffwrf ${WRF_BIN_DIR}/diffwrf_nc
-
-  /bin/mv -fv tools/fseeko_test ${WRF_BIN_DIR}
-  /bin/mv -fv tools/registry ${WRF_BIN_DIR}
-
+  ${ROMS_SRC_DIR}/ESM/wrf_move.csh
 endif
 
 #--------------------------------------------------------------------------
-# Create WRF data links.
+# Create WRF data links to working project directory.
 #--------------------------------------------------------------------------
 
-if ( $WRF_CASE == "em_real" ) then
-
-  echo ""
-  echo "${separator}"
-  echo "Creating WRF data links:  Case em_real"
-  echo "${separator}"
-  echo ""
-
-  cd ${MY_PROJECT_DIR}
-
-  find ./ -type l -exec /bin/rm -f {} \;
-
-  ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA.expanded_rain .
-  ln -sfv ${WRF_ROOT_DIR}/run/RRTM_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_LW_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_SW_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAM_ABS_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAM_AEROPT_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAMtr_volume_mixing_ratio.RCP4.5 .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAMtr_volume_mixing_ratio.RCP6 .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAMtr_volume_mixing_ratio.RCP8.5 CAMtr_volume_mixing_ratio
-  ln -sfv ${WRF_ROOT_DIR}/run/CAMtr_volume_mixing_ratio.A1B .
-  ln -sfv ${WRF_ROOT_DIR}/run/CAMtr_volume_mixing_ratio.A2 .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_ALB_ICE_DFS_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_ALB_ICE_DRC_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_ASM_ICE_DFS_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_ASM_ICE_DRC_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_DRDSDT0_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_EXT_ICE_DFS_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_EXT_ICE_DRC_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_KAPPA_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/CLM_TAU_DATA .
-  ln -sfv ${WRF_ROOT_DIR}/run/ozone.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/ozone_lat.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/ozone_plev.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/aerosol.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/aerosol_lat.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/aerosol_lon.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/aerosol_plev.formatted .
-  ln -sfv ${WRF_ROOT_DIR}/run/capacity.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/coeff_p.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/coeff_q.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/constants.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/masses.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/termvels.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/kernels.asc_s_0_03_0_9 .
-  ln -sfv ${WRF_ROOT_DIR}/run/kernels_z.asc .
-  ln -sfv ${WRF_ROOT_DIR}/run/bulkdens.asc_s_0_03_0_9 .
-  ln -sfv ${WRF_ROOT_DIR}/run/bulkradii.asc_s_0_03_0_9 .
-  ln -sfv ${WRF_ROOT_DIR}/run/CCN_ACTIVATE.BIN .
-  if ( $WRF_VERSION == "4.3" ) then
-    ln -sfv ${WRF_ROOT_DIR}/run/p3_lookupTable_1.dat-2momI_v5.1.6_oldDimax .
-    ln -sfv ${WRF_ROOT_DIR}/run/p3_lookupTable_1.dat-3momI_v5.1.6 .
-    ln -sfv ${WRF_ROOT_DIR}/run/p3_lookupTable_2.dat-4.1 .
-  else
-    ln -sfv ${WRF_ROOT_DIR}/run/p3_lookup_table_1.dat-v4.1 .
-    ln -sfv ${WRF_ROOT_DIR}/run/p3_lookup_table_2.dat-v4.1 .
-  endif
-
-  if ( $?USE_REAL_DOUBLE ) then
-    ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA_DBL ETAMPNEW_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA.expanded_rain_DBL ETAMPNEW_DATA.expanded_rain
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTM_DATA_DBL RRTM_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_LW_DATA_DBL RRTMG_LW_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_SW_DATA_DBL RRTMG_SW_DATA
-  else
-    ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA ETAMPNEW_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/ETAMPNEW_DATA.expanded_rain ETAMPNEW_DATA.expanded_rain
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTM_DATA RRTM_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_LW_DATA RRTMG_LW_DATA
-    ln -sfv ${WRF_ROOT_DIR}/run/RRTMG_SW_DATA RRTMG_SW_DATA
-  endif
-
-  ln -sfv ${WRF_ROOT_DIR}/run/GENPARM.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/LANDUSE.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/SOILPARM.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/URBPARM.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/VEGPARM.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/MPTABLE.TBL .
-  ln -sfv ${WRF_ROOT_DIR}/run/tr49t67 .
-  ln -sfv ${WRF_ROOT_DIR}/run/tr49t85 .
-  ln -sfv ${WRF_ROOT_DIR}/run/tr67t85 .
-  ln -sfv ${WRF_ROOT_DIR}/run/gribmap.txt .
-  ln -sfv ${WRF_ROOT_DIR}/run/grib2map.tbl .
-
-# Set directory structure needed to compile WPS.
-
-  echo ""
-  echo "${separator}"
-  echo "Creating WPS directory layout."
-  echo "${separator}"
-  echo ""
-
-  /bin/mkdir -vp ${WRF_BUILD_DIR}/external/io_grib1
-  /bin/mkdir -vp ${WRF_BUILD_DIR}/external/io_grib_share
-  /bin/mkdir -vp ${WRF_BUILD_DIR}/external/io_int
-  /bin/mkdir -vp ${WRF_BUILD_DIR}/external/io_netcdf
-  /bin/mkdir -vp ${WRF_BUILD_DIR}/frame
-
-# We do not copy files out of the inc directory so we create a symlink
-# do that directory.
-
-  ln -sfv ${WRF_ROOT_DIR}/inc .
-
-  # io_grib1
-
-  echo "cd ${WRF_BUILD_DIR}/external/io_grib1"
-  cd ${WRF_BUILD_DIR}/external/io_grib1
-  ln -sfv ../../libio_grib1.a .
-
-  # io_grib_share
-
-  echo "cd ${WRF_BUILD_DIR}/external/io_grib_share"
-  cd ${WRF_BUILD_DIR}/external/io_grib_share
-  ln -sfv ../../libio_grib_share.a .
-  ln -sfv ../../wrf_io_flags.h     .
-  ln -sfv ../../wrf_status_codes.h .
-
-  # io_int
-
-  echo "cd ${WRF_BUILD_DIR}/external/io_int"
-  cd ${WRF_BUILD_DIR}/external/io_int
-  ln -sfv ../../libwrfio_int.a                  .
-  ln -sfv ../../module_internal_header_util.mod .
-
-  # io_netcdf
-
-  echo "cd ${WRF_BUILD_DIR}/external/io_netcdf"
-  cd ${WRF_BUILD_DIR}/external/io_netcdf
-  ln -sfv ../../libwrfio_nf.a .
-
-  # frame
-
-  echo "cd ${WRF_BUILD_DIR}/frame"
-  cd ${WRF_BUILD_DIR}/frame
-  ln -sfv ../pack_utils.o                  .
-  ln -sfv ../module_machine.o              .
-  ln -sfv ../module_internal_header_util.o .
-  ln -sfv ../module_driver_constants.o     .
-
-# Remove links in WRF/test/em_real sub-directory
-
-  echo ""
-  echo "${separator}"
-  echo "Removing WRF data links from  ${WRF_ROOT_DIR}/test/em_real:"
-  echo "${separator}"
-  echo ""
-
-  find ${WRF_ROOT_DIR}/test/em_real -type l -exec /bin/rm -fv {} \;
-
+if ( $move == 1 ) then
+ if ( $WRF_CASE == "em_real" ) then
+  ${ROMS_SRC_DIR}/ESM/wrf_links.csh
+ endif
 endif
